requirements.txt
click==8.1.7
PyYAML==6.0.1
pip_audit==2.7.3
flake8==7.1.1
coverage==7.8.0
flake8-bugbear==24.12.12
